tts-service/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# PyTorch (CUDA 12.1)
torch==2.1.0
torchaudio==2.1.0

# Audio processing
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1

# Utilities
python-dotenv==1.0.0
